requirements.txt
streamlit>=1.35.0
pdfplumber>=0.11.0
python-docx>=1.1.2
openai>=1.30.0
faiss-cpu>=1.8.0
numpy>=1.26.0
